Tidy structure: docs to docs/, drop empty docker dir, dead test/header; move preprocess_cuda.h to include/snappy_cpp
