+++ src/snappy_cpp/src/camera_inference.cpp
@@ -74,7 +74,7 @@
 #include "snappy_cpp/msg/object_detection.hpp"
 #include "snappy_cpp/msg/bounding_box2_d.hpp"
 
-#include "preprocess_cuda.h"
+#include "snappy_cpp/preprocess_cuda.h"
 
 using namespace std::chrono_literals;
