+++ src/snappy_cpp/src/preprocess_cuda.cu
@@ -1,7 +1,7 @@
 // Fused letterbox + BGR->RGB + normalize + HWC->CHW preprocessing kernel.
-// See preprocess_cuda.h for the contract.
+// See snappy_cpp/preprocess_cuda.h for the contract.
 
-#include "preprocess_cuda.h"
+#include "snappy_cpp/preprocess_cuda.h"
 
 namespace {
